Remove indepedent tls dependencies and their tox environments.
Based on my comment here
https://github.com/twisted/treq/pull/175#issuecomment-289319932:

I'm a little confused by the purpose of the
twisted_latest-pyopenssl_lowest environments. Twisted itself requires
a specific version of pyOpenSSL that's liable to be newer than the
lowest version treq claims to support. Indeed that's the case with
Twisted 17.1.0, which is why this environment fails:

https://travis-ci.org/twisted/treq/jobs/215313795

OP_SINGLE_ECDH_USE was introduced in pyOpenSSL 16.0.0, which Twisted
17.1.0 depends on. So it's no surprise that installing an older
version of pyOpenSSL will break things and it doesn't seem valuable to
test.

Why does treq need to depend on pyOpenSSL (or service_identity) at
all? Depending on Twisted with the tls extra installs the correct
versions of both. Removing a version specification for these
dependencies from treq's setup.py and the related tox environments
would make builds more robust and more honestly represent the actual
dependencies needed.