Run local httpbin processes for integration tests.
https://github.com/mopemope/meinheld/issues/83 means that httpbin.org
and treq (via Twisted) disagree on duplicate header values.
https://tools.ietf.org/html/rfc7230#section-3.2.2 allows servers both
combine repeated header key-value pairs into a single key, values list
and to not do so.

This commit sidesteps that issue and also eliminates an external
network dependency.  httpbin is run in separate processes allows the
tests not to trust httpbin to allow WSGI threads to shut down cleanly.