Patch treq so public APIs don't swallow any unexpected kwargs.
Up until now it was easy to, for example mistype `unbuffered=True` as
`unbuffreed=True` which could have dare consequences :-).

In general it sounds like complaining about unexpected data is a safer bet than
just ignoring it.
Also: Documented the files argument to `treq.request` which was only documented
in the actual code for `treq.client.HTTPClient.request`.